Validation, Swagger, log4j

+++ library/src/main/java/ru/itgirls/library/service/AuthorServiceImpl.java
@@ -1,11 +1,9 @@
 package ru.itgirls.library.service;
 
-import jakarta.persistence.criteria.CriteriaBuilder;
-import jakarta.persistence.criteria.CriteriaQuery;
-import jakarta.persistence.criteria.Predicate;
-import jakarta.persistence.criteria.Root;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 import ru.itgirls.library.dto.AuthorCreateDto;
 import ru.itgirls.library.dto.AuthorDto;
@@ -13,11 +11,13 @@ import ru.itgirls.library.dto.AuthorUpdateDto;
 import ru.itgirls.library.dto.BookDto;
 import ru.itgirls.library.model.Author;
 import ru.itgirls.library.repository.AuthorRepository;
-
 import java.util.Collections;
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class AuthorServiceImpl implements AuthorService {
@@ -26,8 +26,16 @@ public class AuthorServiceImpl implements AuthorService {
 
     @Override
     public AuthorDto getAuthorById(Long id) {
-        Author author = authorRepository.findById(id).orElseThrow();
-        return convertToDto(author);
+        log.info("Try to find author by id {}", id);
+        Optional<Author> author = authorRepository.findById(id);
+        if (author.isPresent()) {
+            AuthorDto authorDto = convertToDto(author.get());
+            log.info("Author: {}", authorDto.toString());
+            return authorDto;
+        } else {
+            log.error("Author with id {} hadn't been found by id.", id);
+            throw new NoSuchElementException("No value present.");
+        }
     }
 
     private AuthorDto convertToDto(Author author) {
@@ -48,50 +56,83 @@ public class AuthorServiceImpl implements AuthorService {
     }
 
     @Override
-    public List<AuthorDto> getAuthorsByName(String name) {
-        Specification<Author> specification = Specification.where(new Specification<Author>() {
-            @Override
-            public Predicate toPredicate(Root<Author> root,
-                                         CriteriaQuery<?> query,
-                                         CriteriaBuilder criteriaBuilder) {
-                return criteriaBuilder.equal(root.get("name"), name);
+    public List<AuthorDto> getAuthorsByName(@NonNull String name) {
+        log.info("Try to find author by name {}", name);
+        try {
+            Specification<Author> specification = (root, query, criteriaBuilder) ->
+                    criteriaBuilder.equal(root.get("name"), name);
+
+            List<Author> authors = authorRepository.findAll(specification);
+            if (authors.isEmpty()) {
+                log.warn("Authors with name {} hadn't been found.", name);
+                return Collections.emptyList();
             }
-        });
-        List<Author> authors = authorRepository.findAll(specification);
-        if (authors.isEmpty()) {
+            log.info("Authors with name {} had been found.", name);
+            List<AuthorDto> authorsDto = authors.stream()
+                    .map(this::convertEntityToDto)
+                    .toList();
+            log.info("Authors: {}", authorsDto);
+            return authorsDto;
+        } catch (Exception e) {
+            log.error("Error occurred while searching of authors with name {}: {}", name, e.getMessage(), e);
             return Collections.emptyList();
         }
-        return authors.stream()
-                .map(this::convertEntityToDto)
-                .toList();
     }
 
     @Override
     public AuthorDto createAuthor(AuthorCreateDto authorCreateDto) {
-        Author author = authorRepository.save(convertDtoToEntity(authorCreateDto));
-        AuthorDto authorDto = convertEntityToDto(author);
-        return authorDto;
+        log.info("Create a new author.");
+        try {
+            Author author = authorRepository.save(convertDtoToEntity(authorCreateDto));
+            AuthorDto authorDto = convertEntityToDto(author);
+            log.info("The author with ID {} had been created.", author.getId());
+            return authorDto;
+        } catch (Exception e) {
+            log.error("Error occurred while creating of the author: {}", e.getMessage(), e);
+            throw e;
+        }
     }
 
     @Override
     public AuthorDto updateAuthor(AuthorUpdateDto authorUpdateDto) {
-        Author author = authorRepository.findById(authorUpdateDto.getId()).orElseThrow();
-        author.setName(authorUpdateDto.getName());
-        author.setSurname(authorUpdateDto.getSurname());
-        Author savedAuthor = authorRepository.save(author);
-        AuthorDto authorDto = convertEntityToDto(savedAuthor);
-        return authorDto;
+        log.info("Update the author with ID: {}", authorUpdateDto.getId());
+        Optional<Author> authorOptional = authorRepository.findById(authorUpdateDto.getId());
+        if (authorOptional.isPresent()) {
+            Author author = authorOptional.get();
+            author.setName(authorUpdateDto.getName());
+            author.setSurname(authorUpdateDto.getSurname());
+            Author savedAuthor = authorRepository.save(author);
+            AuthorDto authorDto = convertEntityToDto(savedAuthor);
+            log.info("Author: {} had been updated.", authorDto.toString());
+            return authorDto;
+        } else {
+            log.error("Author with id {} hadn't been found.", authorUpdateDto.getId());
+            throw new NoSuchElementException("No value present.");
+        }
     }
 
     @Override
     public void deleteAuthor(Long id) {
+        log.info("Delete the author with ID: {}", id);
+        if (!authorRepository.existsById(id)) {
+            log.warn("The author with ID {} hadn't been found, deletion aborted.", id);
+            throw new NoSuchElementException("No value present.");
+        }
         authorRepository.deleteById(id);
+        log.info("The author with ID {} had been deleted", id);
     }
 
     @Override
     public List<AuthorDto> getAllAuthors() {
-        List<Author> authors = authorRepository.findAll();
-        return authors.stream().map(this::convertEntityToDto).collect(Collectors.toList());
+        log.info("Try to get all authors.");
+        try {
+            List<Author> authors = authorRepository.findAll();
+            log.info("{} authors had been found.", authors.size());
+            return authors.stream().map(this::convertEntityToDto).collect(Collectors.toList());
+        } catch (Exception e) {
+            log.error("Error occurred while searching authors: {}", e.getMessage(), e);
+            return Collections.emptyList();
+        }
     }
 
     private Author convertDtoToEntity(AuthorCreateDto authorCreateDto) {
@@ -113,13 +154,11 @@ public class AuthorServiceImpl implements AuthorService {
                             .build())
                     .toList();
         }
-
-        AuthorDto authorDto = AuthorDto.builder()
+        return AuthorDto.builder()
                 .id(author.getId())
                 .name(author.getName())
                 .surname(author.getSurname())
                 .books(bookDtoList)
                 .build();
-        return authorDto;
     }
 }
